Break unit tests out into individual files by area.
