--- test/index.js
@@ -1,2205 +1,0 @@
-"use strict";
-/**
- * Module dependencies.
- */
-
-require('mocha-generators').install();
-
-var Nightmare = require('..');
-var IPC = require('../lib/ipc');
-var chai = require('chai');
-var asPromised = require('chai-as-promised');
-var url = require('url');
-var server = require('./server');
-var fs = require('fs');
-var mkdirp = require('mkdirp');
-var path = require('path');
-var PNG = require('pngjs').PNG;
-var rimraf = require('rimraf');
-var split2 = require("split2");
-var child_process = require('child_process');
-
-var should = chai.should();
-var expect = chai.expect;
-
-chai.use(asPromised);
-
-/**
- * Temporary directory
- */
-
-var tmp_dir = path.join(__dirname, 'tmp');
-
-/**
- * Get rid of a warning.
- */
-
-process.setMaxListeners(0);
-
-/**
- * Locals.
- */
-
-var base = 'http://localhost:7500/';
-
-describe('Nightmare', function () {
-    before(function (done) {
-    server.listen(7500, done);
-    Nightmare = withDeprecationTracking(Nightmare);
-    });
-
-    after(function() {
-        Nightmare.assertNoDeprecations();
-    });
-
-    it('should be constructable', function* () {
-        var nightmare = new Nightmare();
-        nightmare.should.be.ok;
-        nightmare.end();
-    });
-
-    it('should have version information', function* () {
-        let nightmare = new Nightmare();
-        let versions = yield nightmare.init();
-        versions.electron.should.be.ok;
-        versions.chrome.should.be.ok;
-        
-        nightmare.engineVersions.electron.should.be.ok;
-        nightmare.engineVersions.chrome.should.be.ok;
-
-        Nightmare.version.should.be.ok;
-        nightmare.end();
-    });
-
-    it('should kill its electron process when it is killed', function (done) {
-        var child = child_process.fork(
-            path.join(__dirname, 'files', 'nightmare-unended.js'));
-
-        child.once('message', function (electronPid) {
-            child.once('exit', function () {
-                try {
-                    electronPid.should.not.be.a.process;
-                }
-                catch (error) {
-                    // if the test failed, clean up the still-running process
-                    process.kill(electronPid, 'SIGINT');
-                    throw error;
-                }
-                done();
-            });
-            child.kill();
-        });
-    });
-
-    it('should exit with a non-zero code on uncaughtExecption', function (done) {
-        var child = child_process.fork(
-            path.join(__dirname, 'files', 'nightmare-error.js'), [], { silent: true });
-
-        child.once('exit', function (code) {
-            code.should.not.equal(0);
-            done();
-        });
-    });
-
-    describe('navigation', function () {
-        var nightmare;
-
-        beforeEach(function* () {
-            nightmare = new Nightmare({
-                webPreferences: {partition: 'test-partition' + Math.random()}
-            });
-            yield nightmare.init();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should return data about the response', function* () {
-          let data = yield nightmare.goto(fixture('navigation'));
-
-          data.should.contain.keys('url', 'code', 'method', 'referrer', 'headers');
-        });
-
-        it('should click on a link and then go back', function* () {
-
-            var title = yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .title();
-
-            title.should.equal('Navigation');
-
-            title = yield nightmare.chain()
-                .clickAndWaitUntilFinishLoad('a')
-                .title();
-
-            title.should.equal('A');
-
-            var title = yield nightmare.chain()
-                .back()
-                .title();
-
-            title.should.equal('Navigation');
-        });
-
-        it('should fail if navigation target is invalid', function* () {
-            let error = {};
-
-            try {
-                yield nightmare.goto('http://this-is-not-a-real-domain.com');
-            }
-            catch(ex) {
-               error = ex; 
-            }
-
-            error.should.contain.keys('message', 'code', 'url');
-            error.code.should.be.a('number');
-        });
-
-        it('should fail if navigation target is a malformed URL', function* () {
-            let error = {};
-
-            try {
-                yield nightmare.goto('somewhere out there');
-            }
-            catch(ex) {
-                error = ex;
-            }
-            
-            error.details.should.equal('ERR_INVALID_URL');
-        });
-
-        it('should fail if navigating to an unknown protocol', function* () {
-          let error = {};
-
-          try {
-            yield nightmare.goto('fake-protocol://blahblahblah');
-          }
-          catch(ex) {
-            error = ex;
-          }
-          error.details.should.equal('ERR_INVALID_URL');
-        });
-
-        it('should not fail if the URL loads but a resource fails', function* () {
-            yield nightmare.goto(fixture('navigation/invalid-image'));
-        });
-
-        it('should not fail if a child frame fails', function* () {
-            yield nightmare.goto(fixture('navigation/invalid-frame'));
-        });
-
-        it('should return correct data when child frames are present', function* () {
-            let data = yield nightmare.goto(fixture('navigation/valid-frame'));
-
-            data.should.have.property('url');
-            data.url.should.equal(fixture('navigation/valid-frame'));
-        });
-
-        it('should not fail if response was a valid error (e.g. 404)', function* () {
-          yield nightmare.goto(fixture('navigation/not-a-real-page'));
-        });
-
-        it('should fail if the response dies in flight', function* () {
-          let error = {};
-
-          try {
-            yield nightmare.goto(fixture('do-not-respond'));
-          }
-          catch(ex) {
-            error = ex;
-          }
-
-          error.code.should.equal(-324);
-        });
-
-        it('should not fail for a redirect', function* () {
-            yield nightmare.goto(fixture('redirect?url=%2Fnavigation'));
-        });
-
-        it('should fail for a redirect to an invalid URL', function* () {
-          var error = {};
-
-          try {
-              yield nightmare.goto(
-                fixture('redirect?url=http%3A%2F%2Fthis-is-not-a-real-domain.com'));
-          }
-          catch(ex) {
-            error = ex;
-          }
-          
-          error.code.should.equal(-105);
-        });
-
-        it('should succeed properly if request handler is present', function* () {
-            Nightmare.action('monitorRequest',
-                function(name, options, parent, win, renderer) {
-                    win.webContents.session.webRequest.onBeforeRequest(['*://localhost:*'],
-                        function(details, callback) {
-                            callback({cancel: false});
-                        }
-                    );
-                },
-                function() {
-                }
-            );
-
-            let nm = new Nightmare({webPreferences: {partition: 'test-partition'}});
-            yield nm.chain()
-                .goto(fixture('navigation'))
-                .end();
-
-        });
-
-        it('should fail properly if request handler is present', function* () {
-            let error = {};
-            Nightmare.action('monitorRequest',
-                function(name, options, parent, win, renderer) {
-                    win.webContents.session.webRequest.onBeforeRequest(['*://localhost:*'],
-                    function(details, callback) {
-                        callback({cancel: false});
-                    }
-                    );
-                },
-                function() {
-                }
-            );
-
-            let nm = new Nightmare({webPreferences: {partition: 'test-partition'}});
-            yield nm.init();
-
-            try {
-                yield nm.chain()
-                        .goto('http://this-is-not-a-real-domain.com');
-
-            }
-            catch(ex) {
-                error = ex;
-            }
-
-            error.code.should.equal(-105);
-        });
-
-        it('should work for links that dont go anywhere', function* () {
-            var title = yield Promise.all([
-                yield nightmare.goto(fixture('navigation')),
-                yield nightmare.clickAndWaitUntilFinishLoad('a'),
-                yield nightmare.title()
-            ]);
-
-            title[2].should.equal('A');
-
-            title = yield Promise.all([
-                yield nightmare.click('.d'),
-                yield nightmare.title()
-            ]);
-
-            title[1].should.equal('A');
-        });
-
-        it('should click on a link, go back, and then go forward', function* () {
-            yield nightmare.goto(fixture('navigation'));
-            var title = yield nightmare.title();
-            title.should.equal('Navigation');
-            yield nightmare.clickAndWaitUntilFinishLoad('a');
-
-            yield nightmare.back();
-            yield nightmare.forward();
-
-            var title = yield nightmare.title();
-            title.should.equal('A');
-        });
-
-        it('should perform an action that expects navigation', function* () {
-            yield nightmare.goto(fixture('navigation'));
-            var title = yield nightmare.title();
-            title.should.equal('Navigation');
-            yield nightmare.expectNavigation(function() { return this.click('a'); }, 2000);
-
-            var title = yield nightmare.title();
-            title.should.equal('A');
-        });
-
-        it('should fail when an action that expects navigation exceeds the timeout', function* () {
-            var didFail = false;
-            try {
-                yield nightmare.goto(fixture('navigation'));
-                var title = yield nightmare.title();
-                title.should.equal('Navigation');
-                yield nightmare.expectNavigation(function () { return; }, 1000);
-            }
-            catch (ex) {
-                didFail = true;
-            }
-
-            didFail.should.be.true;
-        });
-
-        it('should refresh the page', function* () {
-            yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .refresh();
-        });
-
-        it('should wait until element is present', function* () {
-            yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait('a');
-        });
-
-        it('should escape the css selector correctly when waiting for an element', function* () {
-            yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait('#escaping\\:test');
-        });
-
-        it('should wait until the evaluate fn returns true', function* () {
-            yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait(function () {
-                    var text = document.querySelector('a').textContent;
-                    return (text === 'A');
-                });
-        });
-
-        it('should wait until the evaluate fn with arguments returns true', function* () {
-            yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait(function (expectedA, expectedB) {
-                    var textA = document.querySelector('a.a').textContent;
-                    var textB = document.querySelector('a.b').textContent;
-                    return (expectedA === textA && expectedB === textB);
-                }, 'A', 'B');
-        });
-
-        it('should support javascript URLs', function* () {
-            var gotoResult = yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .goto('javascript:void(document.querySelector(".a").textContent="LINK");');
-            gotoResult.should.be.an('object');
-
-            var linkText = yield nightmare
-                .evaluate(function () {
-                    return document.querySelector('.a').textContent;
-                });
-            linkText.should.equal('LINK');
-        });
-
-        it('should support javascript URLs that load pages', function* () {
-            var data = yield nightmare.chain()
-                .goto(fixture('navigation'))
-                .goto(`javascript:window.location='${fixture('navigation/a.html')}'`);
-            data.should.contain.keys('url', 'code', 'method', 'referrer', 'headers');
-            data.url.should.equal(fixture('navigation/a.html'));
-
-            var linkText = yield nightmare
-                .evaluate(function () {
-                    return document.querySelector('.d').textContent;
-                });
-            linkText.should.equal('D');
-        });
-    });
-
-    describe('evaluation', function () {
-        var nightmare;
-
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should get the title', function* () {
-            var title = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .title();
-            title.should.eql('Evaluation');
-        });
-
-        it('should get the url', function* () {
-            var url = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .url();
-            url.should.have.string(fixture('evaluation'));
-        });
-
-        it('should check if the selector exists', function* () {
-            // existent element
-            var exists = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .exists('h1.title');
-            
-            exists.should.be.true;
-
-            // non-existent element
-            exists = yield nightmare.exists('a.blahblahblah');
-            exists.should.be.false;
-        });
-
-        it('should check if an element is visible', function* () {
-            // visible element
-            var visible = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .visible('h1.title');
-            visible.should.be.true;
-
-            // hidden element
-            visible = yield nightmare
-                .visible('.hidden');
-            visible.should.be.false;
-
-            // non-existent element
-            visible = yield nightmare
-                .visible('#asdfasdfasdf');
-            visible.should.be.false;
-        });
-
-        it('should evaluate javascript on the page, with parameters', function* () {
-            var title = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .evaluate(function (parameter) {
-                    return document.title + ' -- ' + parameter;
-                }, 'testparameter');
-            title.should.equal('Evaluation -- testparameter');
-        });
-
-        it('should capture invalid evaluate fn', function () {
-            return nightmare.chain()
-                .goto(fixture('evaluation'))
-                .evaluate('not_a_function')
-                .should.be.rejected;
-        });
-
-        it('should evaluate javascript and return undefined', function* () {
-            var result = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .evaluate(function () {
-                    return undefined;
-                });
-            expect(result).to.be.undefined;
-        });
-
-        it('should evaluate javascript and reject the promise on errors', function* () {
-            try {
-                var result = yield nightmare.chain()
-                    .goto(fixture('evaluation'))
-                    .evaluate(function () {
-                        throw new Error("Catastrophic Error!");
-                    });
-            }
-            catch (ex) {
-                expect(ex).to.equal("Catastrophic Error!");
-            }
-        });
-
-        it('should evaluate async javascript by waiting for promises', function* () {
-            var timeStart = new Date();
-
-            var title = yield nightmare.chain()
-                .goto(fixture('evaluation'))
-                .evaluateAsync(function (parameter) {
-                    var p = new Promise(function (resolve, reject) {
-                        setTimeout(function () {
-                            resolve(document.title + ' -- ' + parameter);
-                        }, 2000);
-                    });
-                    return p;
-                }, 'testparameter');
-            
-            title.should.equal('Evaluation -- testparameter');
-            var diff = new Date() - timeStart;
-            diff.should.be.at.least(2000);
-        });
-
-        it('should evaluate async javascript promise rejections', function* () {
-            var timeStart = new Date();
-
-            var failed = "";
-            try {
-                var title = yield nightmare.chain()
-                    .goto(fixture('evaluation'))
-                    .evaluateAsync(function (parameter) {
-                        var p = new Promise(function (resolve, reject) {
-                            setTimeout(function () {
-                                reject("Catastrophic Error");
-                            }, 2000);
-                        });
-                        return p;
-                    }, 'testparameter');
-            } catch (ex) {
-                failed = ex;
-            }
-
-            failed.should.equal('Catastrophic Error');
-            var diff = new Date() - timeStart;
-            diff.should.be.at.least(2000);
-        });
-
-        it('should evaluate multiple async calls and return proper results', function* () {
-            var timeStart = new Date();
-
-            yield nightmare.chain()
-                .goto(fixture('evaluation'));
-
-            var queue = [];
-            for (let i = 0; i < 3; i++) {
-                var p = nightmare.evaluate(function (ix) {
-                    return ix;
-                }, i);
-                queue.push(p);
-            }
-
-            var result = yield Promise.all(queue);
-
-            result.should.deep.equal([0, 1, 2]);
-        });
-    });
-
-    describe('manipulation', function () {
-        var nightmare;
-
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should inject javascript onto the page', function* () {
-            var globalNumber = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .inject('js', 'test/files/globals.js')
-                .evaluate(function () {
-                    return globalNumber;
-                });
-            globalNumber.should.equal(7);
-
-            var numAnchors = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .inject('js', 'test/files/jquery-2.1.1.min.js')
-                .evaluate(function () {
-                    return $('h1').length;
-                });
-            numAnchors.should.equal(1);
-        });
-
-        it('should inject css onto the page', function* () {
-            var color = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .inject('js', 'test/files/jquery-2.1.1.min.js')
-                .inject('css', 'test/files/test.css')
-                .evaluate(function () {
-                    return $('body').css('background-color');
-                });
-            color.should.equal('rgb(255, 0, 0)');
-        });
-
-        it('should not inject unsupported types onto the page', function* () {
-            try {
-                var color = yield nightmare.chain()
-                    .goto(fixture('manipulation'))
-                    .inject('js', 'test/files/jquery-2.1.1.min.js')
-                    .inject('pdf', 'test/files/test.css')
-                    .evaluate(function () {
-                        return $('body').css('background-color');
-                    });
-
-                color.should.not.equal('rgb(255, 0, 0)');
-            }
-            catch (ex) {
-                //naught.
-            }
-        });
-
-        it('should type', function* () {
-            var input = 'nightmare';
-            var events = input.length * 3;
-
-            var value = yield nightmare.chain()
-                .on('console', function (type, input, message) {
-                    if (type === 'log') events--;
-                })
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', input)
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('nightmare');
-            events.should.equal(0);
-        });
-
-        it('should emulate keystrokes', function* () {
-            var input = 'nightmare';
-            var events = input.length * 3;
-
-            var value = yield nightmare.chain()
-                .on('console', function (type, input, message) {
-                    if (type === 'log') events--;
-                })
-                .goto(fixture('manipulation'))
-                .emulateKeystrokes('input[type=search]', input)
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('nightmare');
-            events.should.equal(0);
-        });
-
-        it('should type integer', function* () {
-            var input = 10;
-            var events = input.toString().length * 3;
-
-            var value = yield nightmare.chain()
-                .on('console', function (type, input, message) {
-                    if (type === 'log') events--;
-                })
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', input)
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('10');
-            events.should.equal(0);
-        });
-
-
-        it('should type object', function* () {
-            var input = {
-                foo: 'bar'
-            };
-            var events = input.toString().length * 3;
-
-            var value = yield nightmare.chain()
-                .on('console', function (type, input, message) {
-                    if (type === 'log') events--;
-                })
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', input)
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('[object Object]');
-            events.should.equal(0);
-        });
-
-        it('should clear inputs', function* () {
-            var input = 'nightmare';
-            var events = input.length * 3;
-
-            var value = yield nightmare.chain()
-                .on('console', function (type, input, message) {
-                    if (type === 'log') events--;
-                })
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', input)
-                .type('input[type=search]')
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('');
-            events.should.equal(0);
-        });
-
-        it('should support inserting text', function* () {
-            var input = 'nightmare insert typing';
-
-            var value = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .insert('input[type=search]', input)
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('nightmare insert typing');
-        });
-
-        it('should support clearing inserted text', function* () {
-
-            var value = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .insert('input[type=search]')
-                .evaluate(function () {
-                    return document.querySelector('input[type=search]').value;
-                });
-
-            value.should.equal('');
-        });
-
-        it('should not type in a nonexistent selector', function () {
-            return nightmare.chain()
-                .goto(fixture('manipulation'))
-                .type('does-not-exist', 'nightmare')
-                .should.be.rejected;
-        });
-
-        it('should not insert in a nonexistent selector', function () {
-            return nightmare.chain()
-                .goto(fixture('manipulation'))
-                .insert('does-not-exist', 'nightmare')
-                .should.be.rejected;
-        });
-
-        it('should blur the active element when something is clicked', function* () {
-            var isBody = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', 'test')
-                .click('p')
-                .evaluate(function () {
-                    return document.activeElement === document.body;
-                });
-            isBody.should.be.true;
-        });
-
-        it('should type and click', function* () {
-            var title = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', 'nightmare')
-                .clickAndWaitUntilFinishLoad('button[type=submit]')
-                .title();
-
-            title.should.equal('Manipulation - Results');
-        });
-
-        it('should type and click several times', function* () {
-            var title = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', 'github nightmare')
-                .clickAndWaitUntilFinishLoad('button[type=submit]')
-                .clickAndWaitUntilFinishLoad('a')
-                .title();
-            title.should.equal('Manipulation - Result - Nightmare');
-        });
-
-        it('should type and emulate a click several times', function* () {
-            var title = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .type('input[type=search]', 'github nightmare')
-                .emulateClick('button[type=submit]')
-                .waitUntilFinishLoad()
-                .emulateClick('a')
-                .waitUntilFinishLoad()
-                .title();
-            title.should.equal('Manipulation - Result - Nightmare');
-        });
-
-        it('should checkbox', function* () {
-            var checkbox = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .check('input[type=checkbox]')
-                .evaluate(function () {
-                    return document.querySelector('input[type=checkbox]').checked;
-                });
-            checkbox.should.be.true;
-        });
-
-        it('should uncheck', function* () {
-            var checkbox = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .check('input[type=checkbox]')
-                .uncheck('input[type=checkbox]')
-                .evaluate(function () {
-                    return document.querySelector('input[type=checkbox]').checked;
-                });
-            checkbox.should.be.false;
-        });
-
-        it('should select', function* () {
-            var select = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .select('select', 'b')
-                .evaluate(function () {
-                    return document.querySelector('select').value;
-                });
-            select.should.equal('b');
-        });
-
-        it('should scroll to specified position', function* () {
-            // start at the top
-            var coordinates = yield nightmare.chain()
-                .viewport(320, 320)
-                .goto(fixture('manipulation'))
-                .evaluate(function () {
-                    return {
-                        top: document.body.scrollTop,
-                        left: document.body.scrollLeft
-                    };
-                });
-            coordinates.top.should.equal(0);
-            coordinates.left.should.equal(0);
-
-            // scroll down a bit
-            coordinates = yield nightmare.chain()
-                .scrollTo(100, 50)
-                .evaluate(function () {
-                    return {
-                        top: document.body.scrollTop,
-                        left: document.body.scrollLeft
-                    };
-                });
-            coordinates.top.should.equal(100);
-            coordinates.left.should.equal(50);
-        });
-
-        it('should scroll to specified selector', function* () {
-            var selector = 'input';
-            // Get actual element coordinates
-            var elemCoordinates = yield nightmare.chain()
-                .viewport(320, 320)
-                .goto(fixture('manipulation'))
-                .evaluate(function (selector) {
-                    var element = document.querySelector(selector);
-                    var rect = element.getBoundingClientRect();
-                    return {
-                        top: Math.round(rect.top),
-                        left: Math.round(rect.left)
-                    };
-                }, selector);
-            elemCoordinates.should.have.property('top');
-            elemCoordinates.top.should.not.be.equal(0);
-            elemCoordinates.should.have.property('left');
-            elemCoordinates.left.should.not.be.equal(0);
-
-            // Scroll to the element
-            var coordinates = yield nightmare.chain()
-                .scrollTo(selector)
-                .evaluate(function () {
-                    return {
-                        top: document.body.scrollTop,
-                        left: document.body.scrollLeft
-                    };
-                });
-            coordinates.top.should.equal(elemCoordinates.top);
-            // TODO: fix this in the fixture
-            // coordinates.left.should.equal(elemCoordinates.left);
-        });
-
-        it('should hover over an element', function* () {
-            var color = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .mouseover('h1')
-                .evaluate(function () {
-                    var element = document.querySelector('h1');
-                    return element.style.background;
-                });
-            color.should.equal('rgb(102, 255, 102)');
-        });
-
-        it('should mousedown on an element', function* () {
-            var color = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .mousedown('h1')
-                .evaluate(function () {
-                    var element = document.querySelector('h1');
-                    return element.style.background;
-                });
-            color.should.equal('rgb(255, 0, 0)');
-        });
-    });
-
-    describe('cookies', function () {
-        var nightmare;
-
-        beforeEach(function* () {
-            nightmare = new Nightmare({
-                webPreferences: { partition: 'test-partition' }
-            });
-
-            yield nightmare.chain().goto(fixture('cookie'));
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('.set(name, value) & .get(name)', function* () {
-            var cookies = nightmare.cookies;
-
-            yield cookies.set('hi', 'hello');
-            var cookie = yield cookies.get('hi');
-
-            cookie.name.should.equal('hi');
-            cookie.value.should.equal('hello');
-            cookie.path.should.equal('/');
-            cookie.secure.should.equal(false);
-        });
-
-        it('.set(obj) & .get(name)', function* () {
-            var cookies = nightmare.cookies;
-
-            yield cookies.set({
-                name: 'nightmare',
-                value: 'rocks',
-                path: '/cookie'
-            });
-            var cookie = yield cookies.get('nightmare');
-
-            cookie.name.should.equal('nightmare');
-            cookie.value.should.equal('rocks');
-            cookie.path.should.equal('/cookie');
-            cookie.secure.should.equal(false);
-        });
-
-        it('.set([cookie1, cookie2]) & .get()', function* () {
-            var cookies = nightmare.cookies;
-            yield cookies.set([
-                {
-                    name: 'hi',
-                    value: 'hello',
-                    path: '/'
-                },
-                {
-                    name: 'nightmare',
-                    value: 'rocks',
-                    path: '/cookie'
-                }
-            ]);
-
-            var cookies = yield cookies.get();
-            cookies.length.should.equal(2);
-
-            // sort in case they come in a different order
-            cookies = cookies.sort(function (a, b) {
-                if (a.name > b.name) return 1;
-                if (a.name < b.name) return -1;
-                return 0;
-            });
-
-            cookies[0].name.should.equal('hi');
-            cookies[0].value.should.equal('hello');
-            cookies[0].path.should.equal('/');
-            cookies[0].secure.should.equal(false);
-
-            cookies[1].name.should.equal('nightmare');
-            cookies[1].value.should.equal('rocks');
-            cookies[1].path.should.equal('/cookie');
-            cookies[1].secure.should.equal(false);
-        });
-
-        it('.set([cookie1, cookie2]) & .get(query)', function* () {
-            var cookies = nightmare.cookies;
-
-            yield cookies.set([
-                {
-                    name: 'hi',
-                    value: 'hello',
-                    path: '/'
-                },
-                {
-                    name: 'nightmare',
-                    value: 'rocks',
-                    path: '/cookie'
-                }
-            ]);
-
-            var cookies = yield cookies.get({ path: '/cookie' });
-            cookies.length.should.equal(1);
-
-            cookies[0].name.should.equal('nightmare');
-            cookies[0].value.should.equal('rocks');
-            cookies[0].path.should.equal('/cookie');
-            cookies[0].secure.should.equal(false);
-        });
-
-        it('.set([cookie]) & .clear(name) & .get(query)', function* () {
-            var cookies = nightmare.cookies;
-
-            yield cookies.set([
-                {
-                    name: 'hi',
-                    value: 'hello',
-                    path: '/'
-                },
-                {
-                    name: 'nightmare',
-                    value: 'rocks',
-                    path: '/cookie'
-                }
-            ]);
-
-            yield cookies.clear('nightmare');
-
-            var cookies = yield cookies.get({ path: '/cookie' });
-
-            cookies.length.should.equal(0);
-        });
-    });
-
-    describe('rendering', function () {
-        var nightmare;
-
-        before(function (done) {
-            mkdirp(tmp_dir, done);
-        });
-
-        after(function (done) {
-            rimraf(tmp_dir, done);
-        });
-
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should take a screenshot', function* () {
-            yield nightmare.chain()
-                .goto(fixture('screenshot'))
-                .screenshot(tmp_dir + '/test.png');
-            var stats = fs.statSync(tmp_dir + '/test.png');
-            stats.size.should.be.at.least(1000);
-        });
-
-        it('should buffer a screenshot', function* () {
-            var image = yield nightmare.chain()
-                .goto(fixture('screenshot'))
-                .screenshot();
-            Buffer.isBuffer(image).should.be.true;
-            image.length.should.be.at.least(1000);
-        });
-
-        it('should take a clipped screenshot', function* () {
-            yield nightmare.chain()
-                .goto(fixture('screenshot'))
-                .screenshot(tmp_dir + '/test-clipped.png', {
-                    x: 200,
-                    y: 100,
-                    width: 100,
-                    height: 100
-                });
-            var statsClipped = fs.statSync(tmp_dir + '/test-clipped.png');
-            statsClipped.size.should.be.at.least(300);
-        });
-
-        it('should buffer a clipped screenshot', function* () {
-            var image = yield nightmare.chain()
-                .goto(fixture('screenshot'))
-                .screenshot({
-                    x: 200,
-                    y: 100,
-                    width: 100,
-                    height: 100
-                });
-            Buffer.isBuffer(image).should.be.true;
-            image.length.should.be.at.least(300);
-        });
-
-        // repeat this test 3 times, since the concern here is non-determinism in
-        // the timing accuracy of screenshots -- it might pass once, but likely not
-        // several times in a row.
-        for (var i = 0; i < 3; i++) {
-            it('should screenshot an up-to-date image of the page (' + i + ')', function* () {
-                var image = yield nightmare.chain()
-                    .goto('about:blank')
-                    .viewport(100, 100)
-                    .evaluate(function () { document.body.style.background = '#900'; })
-                    .evaluate(function () { document.body.style.background = '#090'; })
-                    .screenshot();
-
-                var png = new PNG();
-                var imageData = yield png.parse.bind(png, image);
-                var firstPixel = Array.from(imageData.data.slice(0, 3));
-                firstPixel.should.deep.equal([0, 153, 0]);
-            });
-        }
-
-        it('should screenshot an idle page', function* () {
-            var image = yield nightmare.chain()
-                .goto('about:blank')
-                .viewport(100, 100)
-                .evaluate(function () { document.body.style.background = '#900'; })
-                .evaluate(function () { document.body.style.background = '#090'; })
-                .wait(1000)
-                .screenshot();
-
-            var png = new PNG();
-            var imageData = yield png.parse.bind(png, image);
-            var firstPixel = Array.from(imageData.data.slice(0, 3));
-            firstPixel.should.deep.equal([0, 153, 0]);
-        });
-
-        it('should not subscribe to frames until necessary', function () {
-            var didSubscribe = false;
-            var FrameManager = require('../lib/frame-manager.js');
-            var manager = FrameManager({
-                webContents: {
-                    beginFrameSubscription: function () { didSubscribe = true; },
-                    endFrameSubscription: function () { },
-                    executeJavaScript: function () { }
-                }
-            });
-            didSubscribe.should.be.false;
-        });
-
-        it('should load jquery correctly', function* () {
-            var loaded = yield nightmare.chain()
-                .goto(fixture('rendering'))
-                .wait(3000)
-                .evaluate(function () {
-                    return !!window.jQuery;
-                });
-            loaded.should.be.at.least(true);
-        });
-
-        it('should render fonts correctly', function* () {
-            yield nightmare.chain()
-                .goto(fixture('rendering'))
-                .wait(2000)
-                .screenshot(tmp_dir + '/font-rendering.png');
-            var stats = fs.statSync(tmp_dir + '/font-rendering.png');
-            stats.size.should.be.at.least(1000);
-        });
-
-        it('should save as html', function* () {
-            yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .html(tmp_dir + '/test.html');
-            var stats = fs.statSync(tmp_dir + '/test.html');
-            stats.should.be.ok;
-        });
-
-        it('should render a PDF', function* () {
-            yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .pdf(tmp_dir + '/test.pdf');
-            var stats = fs.statSync(tmp_dir + '/test.pdf');
-            stats.size.should.be.at.least(1000);
-        });
-
-        it('should accept options to render a PDF', function* () {
-            yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .pdf(tmp_dir + '/test2.pdf', { printBackground: false });
-            var stats = fs.statSync(tmp_dir + '/test2.pdf');
-            stats.size.should.be.at.least(1000);
-        });
-
-        it('should return a buffer from a PDF with no path', function* () {
-            var buf = yield nightmare.chain()
-                .goto(fixture('manipulation'))
-                .pdf({ printBackground: false });
-
-            var isBuffer = Buffer.isBuffer(buf);
-
-            buf.length.should.be.at.least(1000);
-            isBuffer.should.be.true;
-        });
-    });
-
-    describe('events', function () {
-        var nightmare;
-
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should fire an event on page load complete', function* () {
-            var fired = false;
-            nightmare
-                .on('did-finish-load', function () {
-                    fired = true;
-                });
-            yield nightmare
-                .goto(fixture('events'));
-            fired.should.be.true;
-        });
-
-        it('should fire an event on javascript error', function* () {
-            var fired = false;
-            nightmare
-                .on('page', function (type, errorMessage, errorStack) {
-                    if (type === 'error') {
-                        fired = true;
-                    }
-                });
-            yield nightmare
-                .goto(fixture('events'));
-            fired.should.be.true;
-        });
-
-        it('should fire an event on javascript console.log', function* () {
-            var log = '';
-
-            nightmare.on('console', function (type, str) {
-                if (type === 'log') log = str;
-            });
-
-            yield nightmare.goto(fixture('events'));
-
-            log.should.equal('my log');
-        });
-
-        it('should fire an event on page load failure', function* () {
-            var fired = false;
-            nightmare
-                .on('did-fail-load', function () {
-                    fired = true;
-                });
-
-            try {
-                yield nightmare
-                    .goto('http://example:port');
-
-            }
-            catch (ex) {
-                //Do Nothing.
-            }
-
-            fired.should.be.true;
-        });
-
-        it('should fire an event on javascript window.alert', function* () {
-            var alert = '';
-            nightmare.on('page', function (type, message) {
-                if (type === 'alert') {
-                    alert = message;
-                }
-            });
-
-            yield nightmare.chain()
-                .goto(fixture('events'))
-                .evaluate(function () {
-                    alert('my alert');
-                });
-            alert.should.equal('my alert');
-        });
-
-        it('should fire an event on javascript window.prompt', function* () {
-            var prompt = '';
-            var response = '';
-            nightmare.on('page', function (type, message, res) {
-                if (type === 'prompt') {
-                    prompt = message;
-                    response = res;
-                }
-            });
-
-            yield nightmare.chain()
-                .goto(fixture('events'))
-                .evaluate(function () {
-                    prompt('my prompt', 'hello!');
-                });
-            prompt.should.equal('my prompt');
-            response.should.equal('hello!');
-        });
-
-        it('should fire an event on javascript window.confirm', function* () {
-            var confirm = '';
-            var response = '';
-            nightmare.on('page', function (type, message, res) {
-                if (type === 'confirm') {
-                    confirm = message;
-                    response = res;
-                }
-            });
-
-            yield nightmare.chain()
-                .goto(fixture('events'))
-                .evaluate(function () {
-                    confirm('my confirm', 'hello!');
-                });
-            confirm.should.equal('my confirm');
-            response.should.equal('hello!');
-        });
-    });
-
-    describe('options', function () {
-        var nightmare;
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should set audio muted state', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            var audioMuted = yield nightmare
-                .setAudioMuted(false);
-            
-            audioMuted.should.eql(false);
-        });
-
-        it('should set useragent', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            var useragent = yield nightmare.chain()
-                .useragent('firefox')
-                .goto(fixture('options'))
-                .evaluate(function () {
-                    return window.navigator.userAgent;
-                });
-            useragent.should.eql('firefox');
-        });
-
-        it('should complete setting useragent', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            yield nightmare.useragent("firefox");
-            yield nightmare.goto(fixture('options'));
-            var useragent = yield nightmare.evaluate(function () {
-                    return window.navigator.userAgent;
-            });
-            
-            useragent.should.eql('firefox');
-        });
-
-        it('should wait and fail with waitTimeout', function () {
-            nightmare = new Nightmare({ waitTimeout: 254 });
-
-            return nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait('foobar')
-                .should.be.rejected;
-        });
-
-        it('should wait and fail with waitTimeout and a ms wait time', function () {
-            nightmare = new Nightmare({ waitTimeout: 254 });
-
-            return nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait(1000)
-                .should.be.rejected;
-        });
-
-        it('should wait and fail with waitTimeout with queued functions', function () {
-            nightmare = new Nightmare({ waitTimeout: 254 });
-
-            return nightmare.chain()
-                .goto(fixture('navigation'))
-                .wait('foobar')
-                .exists('baz')
-                .should.be.rejected;
-        });
-
-        it('should set authentication', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            var data = yield nightmare.chain()
-                .setAuthenticationCredentials('my', 'auth')
-                .goto(fixture('auth'))
-                .evaluate(function () {
-                    return JSON.parse(document.querySelector('pre').innerHTML);
-                });
-          data.should.eql({ name: 'my', pass: 'auth' });
-        });
-
-        it('should set viewport', function* () {
-            var size = { width: 400, height: 300, 'useContentSize': true };
-            nightmare = new Nightmare(size);
-            yield nightmare.init();
-
-            var result = yield nightmare.chain()
-                .goto(fixture('options'))
-                .evaluate(function () {
-                    return {
-                        width: window.innerWidth,
-                        height: window.innerHeight
-                    };
-                });
-            result.width.should.eql(size.width);
-            result.height.should.eql(size.height);
-        });
-
-        it('should set a single header', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            var headers = yield nightmare.chain()
-                .header('X-Nightmare-Header', 'hello world')
-                .goto(fixture('headers'))
-                .evaluate(function () {
-                    return JSON.parse(document.querySelector('pre').innerHTML);
-                });
-            headers['x-nightmare-header'].should.equal('hello world');
-        });
-
-        it('should set all headers', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            var headers = yield nightmare.chain()
-                .header({ 'X-Foo': 'foo', 'X-Bar': 'bar' })
-                .goto(fixture('headers'))
-                .evaluate(function () {
-                    return JSON.parse(document.querySelector('pre').innerHTML);
-                });
-            headers['x-foo'].should.equal('foo');
-            headers['x-bar'].should.equal('bar');
-        });
-
-        it('should set headers for that request', function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-
-            var headers = yield nightmare.chain()
-                .goto(fixture('headers'), { 'X-Nightmare-Header': 'hello world' })
-                .evaluate(function () {
-                    return JSON.parse(document.querySelector('pre').innerHTML);
-                });
-            headers['x-nightmare-header'].should.equal('hello world');
-        });
-
-        it('should allow webPreferences settings', function* () {
-            nightmare = new Nightmare({ webPreferences: { webSecurity: false } });
-            yield nightmare.init();
-
-            var result = yield nightmare.chain()
-                .goto(fixture('options'))
-                .evaluate(function () {
-                    return document.getElementById('example-iframe').contentDocument;
-                });
-
-            result.should.be.ok;
-        });
-
-        it('should be constructable with paths', function* () {
-            nightmare = new Nightmare({ paths: { userData: __dirname } });
-            yield nightmare.init();
-            nightmare.should.be.ok;
-        });
-
-        it('should be constructable with switches', function* () {
-            nightmare = new Nightmare({ switches: {} });
-            yield nightmare.init();
-            nightmare.should.be.ok;
-        });
-
-        it('should allow to use external Electron', function* () {
-            nightmare = new Nightmare({ electronPath: require('electron-prebuilt') });
-            yield nightmare.init();
-            nightmare.should.be.ok;
-        });
-    });
-
-    describe('Nightmare.action(name, fn)', function () {
-        var nightmare;
-
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should support custom actions that are generators', function* () {
-            Nightmare.action("size", function* () {
-                return yield this.evaluate_now(function () {
-                    var w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-                    var h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-                    return {
-                        height: h,
-                        width: w
-                    };
-                });
-            });
-
-            yield nightmare.init();
-
-            var size = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .size();
-
-            size.height.should.be.a('number');
-            size.width.should.be.a('number');
-        });
-
-        it('should support custom actions that are promises', function* () {
-            Nightmare.action("size", function () {
-                return this.evaluate_now(function () {
-                    var w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-                    var h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-                    return {
-                        height: h,
-                        width: w
-                    };
-                });
-            });
-
-            yield nightmare.init();
-
-            var size = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .size();
-
-            size.height.should.be.a('number');
-            size.width.should.be.a('number');
-        });
-
-        it('should support custom actions with arguments', function* () {
-            Nightmare.action("size", function (scale, offset) {
-                return this.evaluate_now(function (scale, offset) {
-                    var w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-                    var h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-                    return {
-                        height: h,
-                        width: w,
-                        scaledHeight: h * scale + offset,
-                        scaledWidth: w * scale + offset
-                    };
-                }, scale, offset);
-            });
-
-            yield nightmare.init();
-
-            var scaleFactor = 2.0;
-            var offsetFactor = 1;
-            
-            var size = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .size(scaleFactor, offsetFactor);
-
-            size.height.should.be.a('number');
-            size.width.should.be.a('number');
-            size.scaledHeight.should.be.a('number');
-            size.scaledWidth.should.be.a('number');
-            size.scaledHeight.should.be.equal(size.height * scaleFactor + offsetFactor);
-            size.scaledWidth.should.be.equal(size.width * scaleFactor + offsetFactor);
-        });
-
-        it('should support custom namespaces', function* () {
-            Nightmare.action("style", {
-                background: function* () {
-                    return yield this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).backgroundColor;
-                    });
-                },
-                color: function* () {
-                    return yield this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).color;
-                    });
-                }
-            });
-
-            yield nightmare.init();
-
-            yield nightmare.goto(fixture('simple'));
-            var background = yield nightmare.style.background();
-            var color = yield nightmare.style.color();
-
-            background.should.equal('rgba(0, 0, 0, 0)');
-            color.should.equal('rgb(0, 0, 0)');
-        });
-
-        it('should support custom namespaces with arguments', function* () {
-            Nightmare.action("math", {
-                multiply: function* (a, b) {
-                    return yield this.evaluate_now(function (a, b) {
-                        return a * b;
-                    }, a, b);
-                }
-            });
-
-            yield nightmare.init();
-
-            yield nightmare.goto(fixture('simple'));
-            var answerToLifeTheUniverseAndEverything = yield nightmare.math.multiply(6, 7);
-
-            answerToLifeTheUniverseAndEverything.should.equal(42);
-        });
-
-        it('should support chaining on custom namespaces', function* () {
-
-            var backgroundCount = 0;
-            var colorCount = 0;
-
-            Nightmare.action("style2", {
-                background: function* () {
-                    backgroundCount++;
-                    return yield this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).backgroundColor;
-                    });
-                },
-                color: function* () {
-                    colorCount++;
-                    return yield this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).color;
-                    });
-                }
-            });
-            
-            var color = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .style2.background()
-                .style2.color();
-
-            color.should.equal('rgb(0, 0, 0)');
-            colorCount.should.equal(1);
-            backgroundCount.should.equal(1);
-        });
-
-        it('should support extending Electron', function* () {
-            Nightmare.action('bind',
-                function (ns, options, parent, win, renderer) {
-                    parent.respondTo('bind', function (name, done) {
-                        "use strict";
-                        if (renderer.listeners(name).length === 0) {
-                            renderer.on(name, function () {
-                                parent.emit.apply(parent, [name].concat(Array.from(arguments).slice(1)));
-                            });
-                        }
-                        done.resolve();
-                    });
-                },
-                function (name, handler) {
-                    var child = this.child;
-                    if (handler) {
-                        child.on(name, handler);
-                    }
-
-                    return child.call ('bind', name);
-                });
-
-            var eventResults;
-
-            yield nightmare.chain()
-                .on('sample-event', function () {
-                    eventResults = arguments;
-                })
-                .goto(fixture('simple'))
-                .bind('sample-event')
-                .evaluate(function () {
-                    this.ipc.send('sample-event', 'sample', 3, {
-                        sample: 'sample'
-                    });
-                });
-
-            eventResults.length.should.equal(3);
-            eventResults[0].should.equal('sample');
-            eventResults[1].should.equal(3);
-            eventResults[2].sample.should.equal('sample');
-        });
-    });
-
-    describe('Nightmare.use', function () {
-        var nightmare;
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-            yield nightmare.init();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should support extending nightmare', function* () {
-            var tagName = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .use(select('h1'));
-
-            tagName.should.equal('H1');
-
-            function select(tagname) {
-                return function* (nightmare) {
-                    return yield this.evaluate(function (tagname) {
-                        return document.querySelector(tagname).tagName;
-                    }, tagname);
-                };
-            }
-        });
-    });
-
-    describe('Nightmare.prototype', function () {
-        var nightmare;
-        beforeEach(function* () {
-            nightmare = new Nightmare();
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should support custom behavior by simply extending the prototype', function* () {
-            Nightmare.prototype.size = function (scale, offset) {
-                return this.evaluate_now(function (scale, offset) {
-                    var w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-                    var h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-                    return {
-                        height: h,
-                        width: w,
-                        scaledHeight: h * scale + offset,
-                        scaledWidth: w * scale + offset
-                    };
-                }, scale, offset);
-            };
-
-            var scaleFactor = 2.0;
-            var offsetFactor = 1;
-
-            var size = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .size(scaleFactor, offsetFactor);
-
-            size.height.should.be.a('number');
-            size.width.should.be.a('number');
-            size.scaledHeight.should.be.a('number');
-            size.scaledWidth.should.be.a('number');
-            size.scaledHeight.should.be.equal(size.height * scaleFactor + offsetFactor);
-            size.scaledWidth.should.be.equal(size.width * scaleFactor + offsetFactor);
-        });
-
-        it('should support custom namespaces by simply extending the prototype', function* () {
-
-            var backgroundCount = 0;
-            var colorCount = 0;
-
-            Nightmare.prototype.MyStyle = class {
-                *background() {
-                    backgroundCount++;
-                    return yield this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).backgroundColor;
-                    });
-                }
-                *color() {
-                    colorCount++;
-                    return yield this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).color;
-                    });
-                }
-            };
-
-            Nightmare.registerNamespace("MyStyle");
-
-            var color = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .MyStyle.background()
-                .MyStyle.color();
-
-            color.should.equal('rgb(0, 0, 0)');
-            colorCount.should.equal(1);
-            backgroundCount.should.equal(1);
-        });
-
-        it('should support custom namespaces by simply extending the prototype with deep this support', function* () {
-
-            Nightmare.prototype.MyStyle2 = class {
-                inner() {
-                    return this.evaluate_now(function () {
-                        return window.getComputedStyle(document.body, null).backgroundColor;
-                    });
-                }
-                outer() {
-                    return this.MyStyle2.inner();
-                }
-            };
-
-            Nightmare.registerNamespace("MyStyle2");
-            
-            let color = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .MyStyle2.outer();
-
-            color.should.equal('rgba(0, 0, 0, 0)');
-        });
-
-        it('should support custom combinations of electron actions and nightmare actions by extending the prototype with arrays', function* () {
-
-            Nightmare.prototype.getTitle = [
-                function (ns, options, parent, win, renderer) {
-                    parent.respondTo('getTitle', function (done) {
-                        done.resolve(win.webContents.getTitle());
-                    });
-                },
-                function () {
-                    return this._invokeRunnerOperation("getTitle");
-                }
-            ];
-
-            yield nightmare.init();
-            yield nightmare.goto(fixture('simple'));
-            let title = yield nightmare.getTitle();
-            title.should.equal('Simple');
-        });
-
-        it('should support custom combinations of electron actions and nightmare actions by extending the prototype with arrays and be chainable', function* () {
-
-            Nightmare.prototype.getTitle = [
-                function (ns, options, parent, win, renderer) {
-                    parent.respondTo('getTitle', function (done) {
-                        done.resolve(win.webContents.getTitle());
-                    });
-                },
-                function () {
-                    return this._invokeRunnerOperation("getTitle");
-                }
-            ];
-
-            let title = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .getTitle();
-
-            title.should.equal('Simple');
-        });
-
-        it('should support custom combinations of electron actions and nightmare actions within namespaces', function* () {
-
-            Nightmare.prototype.MyTitle = (function () { });
-            Nightmare.prototype.MyTitle.prototype.getTitle = [
-                function (ns, options, parent, win, renderer) {
-                    parent.respondTo('getTitle', function (done) {
-                        done.resolve(win.webContents.getTitle());
-                    });
-                },
-                function () {
-                    return this._invokeRunnerOperation("getTitle");
-                }
-            ];
-
-            Nightmare.registerNamespace("MyTitle");
-
-            yield nightmare.init();
-            yield nightmare.goto(fixture('simple'));
-            let title = yield nightmare.MyTitle.getTitle();
-            title.should.equal('Simple');
-        });
-
-        it('should support custom combinations of electron actions and nightmare actions within namespaces and be chainable', function* () {
-
-            Nightmare.prototype.MyTitle = (function () { });
-            Nightmare.prototype.MyTitle.prototype.getTitle = [
-                function (ns, options, parent, win, renderer) {
-                    parent.respondTo('getTitle', function (done) {
-                        done.resolve(win.webContents.getTitle());
-                    });
-                },
-                function () {
-                    return this._invokeRunnerOperation("getTitle");
-                }
-            ];
-
-            Nightmare.registerNamespace("Mytitle");
-            
-            var title = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .MyTitle.getTitle();
-
-            title.should.equal('Simple');
-        });
-        
-         it('should fail if  a namespace is registered twice.', function* () {
-
-            Nightmare.prototype.MyTitle = class {
-                getTitle() {
-                    return this.evaluate_now(function() {
-                        return document.title;
-                    });
-                }
-            };
-
-            var thrown = false;
-            try {
-                Nightmare.action("Mytitle", {
-                    getTitle: function () {
-                        return this.evaluate_now(function() {
-                            return document.title;
-                        });
-                    }
-                });
-                
-                Nightmare.registerNamespace("Mytitle");
-            }
-            catch (ex) {
-                thrown = true;
-            }
-            
-            thrown.should.equal(true);
-        });
-        
-        it('should support namespaces that do not want this binding', function* () {
-                Nightmare.prototype.maths = class {
-                    constructor(nm) {
-                        this._nightmare = nm;
-                    }
-                    get a() {
-                        return this._a;
-                    }
-                    set a(value) {
-                        this._a = value;
-                    }
-                    *getMeaningOfTitle() {
-                        var title = yield this.getTitle();
-                        return title + " " + (this.a + this.b);
-                    }
-                    getTitle() {
-                        return this._nightmare.evaluate_now(function () {
-                            return document.title;
-                        });
-                    }
-                };
-                
-                Nightmare.registerNamespace("maths", false);
-                yield nightmare.init();
-                
-                nightmare.maths.a = 40;
-                nightmare.maths.b = 2;
-                
-                yield nightmare.goto(fixture('simple'));
-               var result =  yield nightmare.maths.getMeaningOfTitle();
-               result.should.equal("Simple 42");
-               nightmare.maths._nightmare.should.equal(nightmare);
-        });
-        
-        it('should support chaining namespaces that do not want this binding', function* () {
-                Nightmare.prototype.moreMaths = class {
-                    constructor(nm) {
-                        this._nightmare = nm;
-                    }
-                    *getMeaningOfTitle() {
-                        var title = yield this.getTitle();
-                        return title + " " + 42;
-                    }
-                    getTitle() {
-                        return this._nightmare.evaluate_now(function () {
-                            return document.title;
-                        });
-                    }
-                };
-                
-                Nightmare.registerNamespace("moreMaths", false);
-
-                var result = yield nightmare.chain()
-                    .goto(fixture('simple'))
-                    .moreMaths.getMeaningOfTitle();
-                result.should.equal("Simple 42");
-        });
-    });
-
-    describe('custom preload script', function () {
-        it('should support passing your own preload script in', function* () {
-            var nightmare = new Nightmare({
-                webPreferences: {
-                    preload: path.join(__dirname, 'fixtures', 'preload', 'index.js')
-                }
-            });
-
-            try {
-                var value = yield nightmare.chain()
-                    .goto(fixture('preload'))
-                    .evaluate(function () {
-                        return window.preload;
-                    });
-                value.should.equal('custom');
-            } finally {
-                nightmare.end();
-            }
-        });
-    });
-
-    describe('devtools', function () {
-        var nightmare;
-
-        beforeEach(function () {
-            Nightmare.action('waitForDevTools',
-                function (ns, options, parent, win, renderer) {
-                    parent.respondTo('waitForDevTools', function (done) {
-                        done.resolve(win.webContents.isDevToolsOpened());
-                    });
-                },
-                function () {
-                    return this._invokeRunnerOperation("waitForDevTools");
-                });
-            nightmare = new Nightmare({ show: true, openDevTools: true });
-
-        });
-
-        afterEach(function* () {
-            nightmare.end();
-        });
-
-        it('should open devtools', function* () {
-            var devToolsOpen = yield nightmare.chain()
-                .goto(fixture('simple'))
-                .wait(1000)
-                .waitForDevTools();
-
-            devToolsOpen.should.be.true;
-        });
-    });
-    
-    describe('ipc', function () {
-        let nightmare;
-        beforeEach(function* () {
-            Nightmare.action('test',
-                function (_, __, parent, ___, ____) {
-                    parent.respondTo('test', function (arg1, done) {
-                        done.progress('one');
-                        done.progress('two');
-                        if (arg1 === 'error') {
-                            done.reject('Error!');
-                        }
-                        else {
-                            done.resolve(`Got ${arg1}`);
-                        }
-                    });
-                },
-                function (options) {
-                    
-                    var promise = this.child.call('test', options.arg || options);
-                    if (options.onData) {
-                        promise.progress.on('data', options.onData);
-                    }
-                    
-                    if (options.onEnd) {
-                        promise.progress.on('end', options.onEnd);
-                    }
-                    
-                    return promise;
-                });
-            Nightmare.action('noImplementation',
-                function () {
-                    return this.child.call('noImplementation');
-                });
-            nightmare = new Nightmare();
-            yield nightmare.init();
-        });
-
-        afterEach(function () {
-            nightmare.end();
-        });
-
-        it('should only make one IPC instance per process', function () {
-            var processStub = { send: function () { }, on: function () { } };
-            var ipc1 = IPC(processStub);
-            var ipc2 = IPC(processStub);
-            ipc1.should.equal(ipc2);
-        });
-
-        it('should support basic call-response', function* () {
-            var result = yield nightmare.test('x');
-            result.should.equal('Got x');
-        });
-
-        it('should support errors across IPC', function (done) {
-            nightmare.test('error').then(
-                function () {
-                    done.reject(new Error('Action succeeded when it should have errored!'));
-                },
-                function () {
-                    done();
-                });
-        });
-
-        it('should stream progress', function* () {
-            var progress = [];
-            yield nightmare.test({
-                arg: 'x',
-                onData: (data) => progress.push(data),
-                onEnd: (data) => progress.push(data)
-            });
-            progress.should.deep.equal(['one', 'two', 'Got x']);
-        });
-
-        it('should trigger error if no responder is registered', function (done) {
-            nightmare.noImplementation().then(
-                function () {
-                    done(new Error('Action succeeded when it should have errored!'));
-                },
-                function () {
-                    done();
-                });
-        });
-
-        it('should log a warning when replacing a responder', function* () {
-            Nightmare.action('uhoh',
-                function (_, __, parent, ___, ____) {
-                    parent.respondTo('test', function (done) {
-                        done.resolve();
-                    });
-                },
-                function () {
-                    return this.child.call('test');
-                });
-
-            let logged = false;
-            let nightmare = new Nightmare();
-
-            var result = yield nightmare.chain({
-                onChildReady: function () {
-                    nightmare.on('nightmare:ipc:debug', function (message) {
-                        if (message.toLowerCase().indexOf('replacing') > -1) {
-                            logged = true;
-                        }
-                    });
-                }
-            })
-                .goto('about:blank')
-                .end();
-                
-            logged.should.be.true;
-        });
-    });
-});
-
-/**
- * Generate a URL to a specific fixture.
- *
- * @param {String} path
- * @returns {String}
- */
-
-function fixture(path) {
-    return url.resolve(base, path);
-}
-
-/**
- * Track deprecation warnings.
- */
-
-function withDeprecationTracking(cls) {
-
-  Nightmare.__deprecations = [];
-
-  cls.prototype.childReady = function() {
-    var self = this;
-    self.proc.stderr.pipe(split2()).on('data', function(line) {
-      if (line.indexOf('deprecated') > -1) {
-        Nightmare.__deprecations.unshift(line);
-      }
-    });
-  };
-  
-  cls.assertNoDeprecations = function() {
-    var deprecations = Nightmare.__deprecations;
-    if (deprecations.length) {
-      var plural = deprecations.length === 1 ? '' : 's';
-      throw new Error(
-        `Used ${deprecations.length} deprecated Electron API${plural}:
-        ${Array.from(deprecations).join('\n        ')}`);
-    }
-  };
-
-  return cls;
-};
-
-/**
- * Simple assertion for running processes
- */
-chai.Assertion.addProperty('process', function () {
-    var running = true;
-    try { process.kill(this._obj, 0); } catch (e) { running = false; }
-    this.assert(
-        running,
-        'expected process ##{this} to be running',
-        'expected process ##{this} not to be running');
-});
